replay viewer: add 1/2x playback speed; pin the board page's Space pause
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) that spends one sim tick every other frame
(halfPhase parity in replayStepBudget, outside the lull boost so a skipped
lull still flies), the chrome grows a 0.5x chip, and the frame JSON's sp
field carries 0.5 while it is active (buildStateJson speed is a float now).
replaySpeed() still clamps to PlaybackSpeeds[0] for every integer consumer,
so the live server loop is unchanged.

client/replay_broadcast.html is the only page this repo ships (there is no
league shell here) and it already bound Space to play/pause; test_viewer now
pins that binding and the digit forwarding that carries '5' to the engine,
so neither can be dropped silently.

chrome_common.js keeps its inherited window.CTF_WIRE read — this repo pins
the fallbacks equal to the engine's wire block by design — so the 0.5 entry
goes into both: the fallback SPEEDS list and WireConstantsJs's emitted
speeds. The viewer test's byte pin moves to the patched file (sha256 +
length), test_server pins the fallback against the wire block, the CI
renderer fixture expects 7 speeds, and the design note gains an erratum.

+++ src/waterworld/broadcast.nim
@@ -265,7 +265,9 @@ proc buildStateJson*(
   sim: SimServer,
   events: JsonNode,
   playing: bool,
-  speed: int,
+  # `speed` is the speed the chrome SHOWS (`sp`), which is fractional at the
+  # replay-only 1/2x — not the engine's integer PlaybackSpeeds value.
+  speed: float,
   maxTick: int,
   looping: bool,
   transportEnabled: bool,
